Remove assumption that there is always a "Username" and "Password" field
Current behaviour tries to pull the "Username" and "Password" fields
out of the data returned from Vault and sticks them in this custom
"Credentials" struct. This is unnessecary and breaks usage in cases
where the returned secret does not have those fields. (For example
with the AWS secrets engine, you get "access_key" and "secret_key").